runtime/funcs: apply review feedback and split core tests by component
